feat: Allow teachers to have the same national ID in different schools
The unique constraint on `national_id` in the `teachers` table has been modified. Previously, it prevented a national ID from being used anywhere in the system. Now, the constraint is applied per school, allowing a teacher to be registered with the same national ID in multiple schools.

This change also includes updated user-facing error messages to reflect the new constraint behavior.